# internal/api/wellknown/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "See docs/10-running-in-production.md for token issuance. A route with no matching scope in the token, and no admin role, is refused with 403.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Every route declared in internal/apispec, the package cmd/controller/main.go also builds its real router from, served under /api/v1. A listed route's method, pattern, scope, and link relation are the ones the server enforces. That a listed route is registered at all is not checked by anything, so confirm a route against a running controller before building on it.",
    "title": "Pleiades control plane API",
    "version": "unreleased"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/announcements": {
      "get": {
        "description": "Returns announcements this caller may see, newest first. System-wide announcements are returned to everybody; tenant-scoped ones only to callers who can reach that tenant. By default only announcements live right now are returned.",
        "operationId": "list_announcements",
        "parameters": [
          {
            "description": "Include announcements outside their active window. For managing them, not for reading them.",
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Maximum announcements to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "announcements": {
                      "items": {
                        "properties": {
                          "_links": {
                            "description": "Hypermedia affordances available on this resource for the calling identity.",
                            "items": {
                              "properties": {
                                "href": {
                                  "type": "string"
                                },
                                "method": {
                                  "type": "string"
                                },
                                "rel": {
                                  "type": "string"
                                }
                              },
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "author": {
                            "description": "Captured at write time and immutable.",
                            "type": "string"
                          },
                          "body": {
                            "description": "The message.",
                            "type": "string"
                          },
                          "ends_at": {
                            "description": "RFC 3339. Absent means until manually retired.",
                            "type": "string"
                          },
                          "id": {
                            "type": "integer"
                          },
                          "level": {
                            "description": "One of info, changed, warning, critical.",
                            "type": "string"
                          },
                          "organization": {
                            "description": "Absent means system-wide, shown to every tenant.",
                            "type": "integer"
                          },
                          "starts_at": {
                            "description": "RFC 3339. Absent means already showing.",
                            "type": "string"
                          },
                          "title": {
                            "description": "The headline.",
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The announcements."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "announcement:read"
            ]
          }
        ],
        "summary": "List operator announcements",
        "x-link-relation": "collection"
      },
      "post": {
        "description": "Puts a message in front of every operator who can see it. The author is taken from the caller's identity and cannot be set.",
        "operationId": "create_announcement",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "author": {
                      "description": "Captured at write time and immutable.",
                      "type": "string"
                    },
                    "body": {
                      "description": "The message.",
                      "type": "string"
                    },
                    "ends_at": {
                      "description": "RFC 3339. Absent means until manually retired.",
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "level": {
                      "description": "One of info, changed, warning, critical.",
                      "type": "string"
                    },
                    "organization": {
                      "description": "Absent means system-wide, shown to every tenant.",
                      "type": "integer"
                    },
                    "starts_at": {
                      "description": "RFC 3339. Absent means already showing.",
                      "type": "string"
                    },
                    "title": {
                      "description": "The headline.",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The created announcement."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Missing title or body."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "announcement:write"
            ]
          }
        ],
        "summary": "Post an announcement",
        "x-link-relation": "create"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the controller's own origin.",
      "url": "/api/v1"
    }
  ]
}